requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dbt-bigquery
astronomer-cosmos
pandas>=3.0.1
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
requests>=2.32.5
sqlalchemy